2020: Communication/Visitation Attempts
All messages have been disregarded although she reads them all. It has been a over a year, with every holiday being denied to see my son. This includes 2 birthdays as of the day of this upload. This type of removal of a child is psychologically abusive and has been the cause of suicidal thoughts, losing work, losing sleep, PTSD, etc...